made CheckVisibility component
