+++ src/components/AwardBox.js
@@ -1,0 +1,21 @@
+import React, {Component} from 'react';
+import checkVisibility from './checkVisibility';
+
+export default class AwardBox extends Component{
+    constructor(){
+        super();
+
+    }
+
+    componentDidMount(){
+    }
+
+    render(){
+        return (
+            <div>
+            </div>
+        )
+    }
+
+}   
+
